Added PayloadQuery. You can now specify which queries will have payload boost, creating an AndQuery with your original query and a PayloadQuery.
Removed payload logic on Searcher. Now the client is responsible of creating a PayloadQuery, the searcher does not do that anymore.

ReloadableIndexHandler now creates a SimilarityForwarder, and that class is responsible of reading PayloadScorers from configuration files.

Added more debug information on FieldFormatCheckerModule.

Added SimilarityForwarder.scorers definition on searcher.properties.defaults. It sends "date_payload" to a DatePayloadScorer (the only one we have right now).