Updated flaptor-util jar. It contains a fix to the ARmiClientStub reconnection mechanism. The interface has changed as well, since now the "checkConnection" method returns a boolean indicating that the call should or should not be done, leaving the exceptions _ONLY_ to indicate true errors.
RmiSearcherStub was modified to comply with the new interface of the checkConnection method of it's parent class (ARmiClientStub).

TrafficLimitingSearcher contained a race condition involving the calculation of the size of the queue. There was the posibility that the newly calculated size of it was seen too late by the enqueuing threads.