Added an "anti-pattern" to the UrlPatterns definitions, so now the urlPatterns are defined as "<prefix> | <pattern-regex> || <anti-pattern-regex> ||| <tokens>". This means that any urlPattern definition file that used tokens will need to change the "||" separator to the "|||" separator.
Added a LoggingSearcher that logs user queries to a rolling file, activated through the compositeSearcher.useLoggingSearcher setting in the searcher.properties file.

Fixed bug in SuggestQuerySearcher that provided suggestions that where equal to or worse than the original query.

Extracted the FindLazyParsedQuery method from AQuerySuggestor for use by other classes.